% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myData.R
\docType{data}
\name{myData}
\alias{myData}
\title{A sample dataset}
\format{
A data frame with 100 rows and 6 variables:
\describe{
\item{Y}{Adjusted student final test scores}
\item{X1}{Adjusted Pre-test score}
\item{X2}{Adjusted Midterm test score}
\item{X3}{Reading level, in relation to number of years above or below current grade level}
\item{X4}{Math level, in relation to number of years above or below current grade level}
\item{X5}{Adjusted Student reported prediction for their final test score}
...
}
}
\source{
Professor Sampson
}
\usage{
myData
}
\description{
A dataset including the final test score for 100 students and and number of associated metrics such as pre-test and midterm scores,Math and reading level in relation to their grade level, and a guess each student made as to what they think their final grade will be.
}
\keyword{datasets}
