% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_hw12.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{My Linear Regression function.}
\usage{
myLinearRegression(x, y, sub)
}
\arguments{
\item{x}{a matrix of  covariates.}

\item{y}{a vector  of outcomes.}

\item{sub}{a list of  subjects (a set of integers corresponding to row numbers in X).}
}
\value{
The coeffcients and p values of \code{x} when \code{y} is the outcome and subset with the rows included in sub.
}
\description{
This function takes in a list of outcomes, a set of covariates, and a list of row numbers. It is important to not that the list of outcomes has the same number of rows as the set of covariates.
The function checks the number of columns and if there are 5 or less in X then a matrix of scatterplots that contain the covariates in X is printed.
The function will then run a linear regression using Y as the outcome and X as the explanatory variables before outputting a table that contains the coefficients and the p-values for those coefficients
Both the linear regression and the matrix of scatterplots will be created with the subset of data indicated in the "sub" list of row numbers.
}
\examples{
myLinearRegression(myData[2:6], myData$Y, c(1:15, 25,27,33,36:45, 51, 55, 57, 59, 63, 66:71, 77, 81, 88, 90:95, 99))
myLinearRegression(Iris[2:4], Iris$Sepal.Length, c(1,3,6,7,8:24, 27, 38, 45:50))
}
